sway: use a very thin title bar instead of the default indicator
the default indicator makes it so hard to know the active window for me.
removing the indicator and using a very thin title bar with a tiny font makes a better indicator in my opinion.

This makes it easy to tell which window is active by following the pink strip rather than associate which window is the one surrounded by pink.